[Fixes #2262] Huge cleanup of files (including removal of akvo.api)
