--- akvo/api/validation.py
@@ -1,62 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Akvo RSR is covered by the GNU Affero General Public License.
-# See more details in the license.txt file located at the root folder of the Akvo RSR module.
-# For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
-
-# original source at https://github.com/toastdriven/django-tastypie/issues/152
-
-from django.forms.models import ModelChoiceField
-
-from tastypie.validation import FormValidation
-
-class ModelFormValidation(FormValidation):
-    """
-    Override tastypie's standard ``FormValidation`` since this does not care
-    about URI to PK conversion for ``ToOneField`` or ``ToManyField``.
-    """
-
-    def uri_to_pk(self, uri):
-
-        """
-        Returns the integer PK part of a URI.
-
-        Assumes ``/api/v1/resource/123/`` format. If conversion fails, this just
-        returns the URI unmodified.
-
-        Also handles lists of URIs
-        """
-
-        if uri is None:
-            return None
-
-        # convert everything to lists
-        multiple = not isinstance(uri, basestring)
-        uris = uri if multiple else [uri]
-
-        # handle all passed URIs
-        converted = []
-        for one_uri in uris:
-            try:
-                # hopefully /api/v1/<resource_name>/<pk>/
-                converted.append(int(one_uri.split('/')[-2]))
-            except (IndexError, ValueError):
-                raise ValueError(
-                    "URI %s could not be converted to PK integer." % one_uri)
-
-        # convert back to original format
-        return converted if multiple else converted[0]
-
-    def form_args(self, bundle):
-        kwargs = super(ModelFormValidation, self).form_args(bundle)
-
-        relation_fields = [name for name, field in
-                           self.form_class.base_fields.items()
-                           if issubclass(field.__class__, ModelChoiceField)]
-
-        for field in relation_fields:
-            if field in kwargs['data']:
-                kwargs['data'][field] = self.uri_to_pk(kwargs['data'][field])
-
-        return kwargs
-
