--- akvo/rest/rsr_api_upload/json/organisation_1.json
@@ -1,30 +1,0 @@
-{
-    "organisation": {
-        "name": "A nice little org",
-        "long_name": "The friendly org with a nice name",
-        "language": "en",
-        "organisation_type": "N",
-        "new_organisation_type": 22,
-        "iati_org_id": "NL-2",
-        "logo": "http://www.flubu.com/blog/wp-content/uploads/2012/07/recharging.jpg",
-        "url": "http://www.exampleorg.com/",
-        "phone": "",
-        "mobile": "",
-        "fax": "",
-        "contact_person": "Spiderman",
-        "contact_email": "ara@chnid.org",
-        "description": "The organisation description",
-        "notes": "",
-        "content_owner": null,
-        "allow_edit": true,
-        "primary_location": null,
-        "internal_org_ids": []
-    },
-    "organisation_location": {
-        "latitude": 59.3293230,
-        "longitude": 18.0685810,
-        "city": "Stockholm",
-        "state": "",
-        "country": 18
-    }
-}
