--- akvo/rest/rsr_api_upload/json/project_1.json
@@ -1,116 +1,0 @@
-{
-    "project": {
-        "title": "API created project",
-        "subtitle": "The subtitle can of course be longer!",
-        "status": "A",
-        "project_plan_summary": "Project plan summary text",
-        "current_image": "json/lolcat.jpg",
-        "current_image_caption": "Optional",
-        "current_image_credit": "Optional",
-        "goals_overview": "Goals overview text",
-        "current_status": "Current status text",
-        "project_plan": "Project plan text",
-        "sustainability": "Sustainability text",
-        "background": "Background text",
-        "target_group": "Target group text",
-        "language": "en",
-        "notes": "",
-        "currency": "EUR",
-        "date_request_posted": "2014-01-01",
-        "date_complete": "2014-06-30",
-        "primary_location": null,
-        "donate_button": true,
-        "categories": [1],
-        "partners": []
-    },
-    "benchmark": [
-        {
-            "category": 1,
-            "name": 1,
-            "value": 50
-        },
-        {
-            "category": 1,
-            "name": 2,
-            "value": 500
-        },
-        {
-            "category": 2,
-            "name": 6,
-            "value": 30
-        }
-    ],
-    "budget_item": [
-        {
-            "label": 6,
-            "amount": "3600.00"
-        },
-        {
-            "label": 5,
-            "amount": "2500.00"
-        },
-        {
-            "label": 1,
-            "other_extra": "Light sabers",
-            "amount": "3995.00"
-        }
-    ],
-    "goal": [
-        {
-            "text": "Use the Force"
-        },
-        {
-            "text": "Ride a Bandersnatch"
-        },
-        {
-            "text": "See the C-beams near Tannhauser gate"
-        }
-    ],
-    "link": {
-        "kind": "E",
-        "url": "https://www.youtube.com/watch?v=_JjJzMBGUwo",
-        "caption": "Roy Batty's last words"
-    },
-    "project_location": {
-        "latitude": 59.3293230,
-        "longitude": 18.0685810,
-        "city": "Stockholm",
-        "state": "",
-        "country": 18
-    },
-    "project_update": {
-        "user": 1,
-        "title": "Project update title",
-        "text": "### A subtitle\n\nSome text",
-        "language": "en",
-        "photo": "http://www.flubu.com/blog/wp-content/uploads/2012/07/recharging.jpg",
-        "photo_caption": "Pic cap",
-        "photo_credit": "Pic cred",
-        "video": "",
-        "video_caption": "",
-        "video_credit": "",
-        "update_method": "W",
-        "user_agent": "",
-        "uuid": "",
-        "notes": "",
-        "locations": []
-    },
-    "partnership": {
-        "organisation": "Akvo",
-        "iati_organisation_role": "4",
-        "funding_amount": null,
-        "partner_type_extra": null,
-        "iati_activity_id": null,
-        "internal_id": null,
-        "iati_url": ""
-    },
-    "partnership": {
-        "organisation": "Akvo",
-        "iati_organisation_role": "1",
-        "funding_amount": 4711,
-        "partner_type_extra": null,
-        "iati_activity_id": null,
-        "internal_id": null,
-        "iati_url": ""
-    }
-}
