+++ akvo/rsr/models/__init__.py
@@ -12,7 +12,7 @@ from django.conf import settings
 from django.db.models.signals import pre_save, post_save, post_delete
 from django.contrib.admin.models import LogEntry
 
-from akvo.api.models import create_api_key
+from akvo.rest.models import create_api_key
 
 from ..signals import (
     change_name_of_file_on_change, change_name_of_file_on_create,
