+++ akvo/settings/10-base.conf
@@ -25,7 +25,6 @@ INSTALLED_APPS = (
     'django.contrib.webdesign',
     'akvo.codelists',
     'akvo.rsr',
-    'akvo.api',
     'registration',
     'template_utils',
     'paypal.standard.ipn',
