+++ akvo/urls.py
@@ -5,7 +5,6 @@ See more details in the license.txt file located at the root folder of the Akvo 
 For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
 """
 
-from akvo.api.urls import named_api
 from akvo.utils import check_auth_groups
 from .rsr.views import widgets as widget_views
